+++ edsnlp/pipes/misc/measurements/measurements.py
@@ -7,6 +7,7 @@ from itertools import repeat
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple, Union
 
 import regex
+from loguru import logger
 from spacy.tokens import Doc, Span
 from typing_extensions import Literal, NotRequired, TypedDict
 
@@ -51,7 +52,7 @@ class UnitlessPatternConfig(TypedDict):
 
 class UnitlessPatternConfigWithName(TypedDict):
     terms: List[str]
-    ranges: List[UnitlessRange]
+    ranges: NotRequired[List[UnitlessRange]]
     name: str
 
 
@@ -108,7 +109,7 @@ class UnitRegistry:
                     "scale": 1 / unit_config["scale"],
                 }
 
-    @lru_cache(maxsize=-1)
+    @lru_cache(maxsize=1000)
     def parse_unit(self, unit: str) -> Tuple[str, float]:
         degrees = defaultdict(lambda: [])
         scale = 1
@@ -270,6 +271,11 @@ class MeasurementsMatcher(BaseNERComponent):
         been rigorously validated. If you come across a measurement expression that
         goes undetected, please file an issue !
 
+    Pipe definition
+    ---------------
+
+    --8<-- "docs/utilities/measurements-utils.md:pipe-definition"
+
     Scope
     -----
     The `eds.measurements` matcher can extract simple (e.g. `3cm`) measurements.
@@ -281,16 +287,7 @@ class MeasurementsMatcher(BaseNERComponent):
     desired unit. Like for other components, the `span._.value` extension can also be
     used to access the normalized value for any measurement span.
 
-
-    The current matcher annotates the following measurements out of the box:
-
-    | Measurement name | Example                |
-    |------------------|------------------------|
-    | `size`           | `1m50`, `1.50m`        |
-    | `weight`         | `12kg`, `1kg300`       |
-    | `bmi`            | `BMI: 24`, `24 kg.m-2` |
-    | `volume`         | `2 cac`, `8ml`         |
-
+    See Availability section for details on which units are handled
 
     Examples
     --------
@@ -427,6 +424,7 @@ class MeasurementsMatcher(BaseNERComponent):
           }
         }
         ```
+        Set `measurements="all"` to extract all measurements from units_config file.
     number_terms: Dict[str, List[str]
         A mapping of numbers to their lexical variants
     stopwords: List[str]
@@ -465,6 +463,11 @@ class MeasurementsMatcher(BaseNERComponent):
           instead, and assign all the parsed information (`._.date` / `._.duration`)
           to it. Otherwise, don't return the date.
 
+    Availability
+    ------------
+    
+    --8<-- "docs/utilities/measurements-utils.md:availability_mes_units"
+
     Authors and citation
     --------------------
     The `eds.measurements` pipeline was developed by AP-HP's Data Science team.
@@ -493,7 +496,23 @@ class MeasurementsMatcher(BaseNERComponent):
           merge_mode: Literal["intersect", "align"] = "intersect",
           as_ents: bool = False,
           span_setter: Optional[SpanSetterArg] = None,
+          use_tables : bool = True
     ):
+
+        self.use_tables = use_tables and (
+            "eds.tables" in nlp.pipe_names or "tables" in nlp.pipe_names
+        )
+        if use_tables and not self.use_tables:
+            logger.warning(
+                "You have requested that the pipeline use annotations "
+                "provided by the `table` pipeline, but it was not set. "
+                "Skipping that step."
+            )
+
+        self.all_measurements = (measurements == "all")
+        if self.all_measurements:
+            measurements = []
+
         # fmt: on
         if isinstance(measurements, str):
             measurements = [measurements]
@@ -556,6 +575,13 @@ class MeasurementsMatcher(BaseNERComponent):
             ignore_excluded=ignore_excluded,
             ignore_space_tokens=True,
         )
+
+        if self.all_measurements:
+            measurements = [
+                {"name": name, **common_measurement}
+                for name, common_measurement in patterns.common_measurements.items()
+            ]
+
         for measure_config in measurements:
             name = measure_config["name"]
             unit = measure_config["unit"]
@@ -788,8 +814,12 @@ class MeasurementsMatcher(BaseNERComponent):
         List[Span]
         """
         doc = doclike.doc if isinstance(doclike, Span) else doclike
-        matches, unit_label_hashes = self.get_matches(doclike)
 
+        table_matches = []
+        if self.use_tables:
+            table_matches = list(doc.spans["tables"])
+
+        matches, unit_label_hashes = self.get_matches(doclike)
         # Make match slice function to query them
         def get_matches_after(i):
             anchor = matches[i][0]
@@ -862,9 +892,44 @@ class MeasurementsMatcher(BaseNERComponent):
             try:
                 pseudo_sent = pseudo[offsets[number_idx] + 1: offsets[unit_idx]]
                 if not re.fullmatch(r"(,n)*", pseudo_sent):
-                    unit_text, unit_norm = None, None
+                    unit_norm = None
             except TypeError:
                 pass
+
+            # Check if number is in table with a unit in the same row
+            if unit_norm is None :
+                for table in table_matches:
+                    if (number.start >= table.start) and (number.end <= table.end):
+                        table_pd = table._.to_pd_table(as_spans=True)
+                        # Find out the number's row
+                        for _, row in table_pd.iterrows():
+                            start_line = next((item.start for item in row
+                                               if item is not None), None)
+                            end_line = next((item.end for item in reversed(row)
+                                             if item is not None), None)
+                            if start_line is None:
+                                continue
+                            def is_within_row(x):
+                                return (x.start >= start_line) and (x.end <= end_line)
+
+                            if is_within_row(number):
+                                # Check if any unit in the same row
+                                if unit_text and is_within_row(unit_text):
+                                    unit_norm = unit_text.label_
+                                    continue
+                                try:
+                                    b_unit_idx, b_unit_text = next(
+                                        (j, ent)
+                                        for j, ent in get_matches_before(number_idx)
+                                        if ent.label in unit_label_hashes
+                                    )
+                                    b_unit_norm = b_unit_text.label_
+                                    if is_within_row(b_unit_text):
+                                        unit_text = b_unit_text
+                                        unit_norm = b_unit_norm
+                                        unit_idx = b_unit_idx
+                                except (AttributeError, StopIteration):
+                                    pass
 
             # Otherwise, try to infer the unit from the preceding unit to handle cases
             # like (1 meter 50)
@@ -878,7 +943,7 @@ class MeasurementsMatcher(BaseNERComponent):
 
             # If no unit was matched, try to detect unitless patterns before
             # the number to handle cases like ("Weight: 63, Height: 170")
-            if not unit_norm:
+            if unit_norm is None:
                 try:
                     (unitless_idx, unitless_text) = next(
                         (j, e)
@@ -930,12 +995,18 @@ class MeasurementsMatcher(BaseNERComponent):
 
             # If the measure was not requested, dismiss it
             # Otherwise, relabel the entity and create the value attribute
-            if dims not in self.measure_names:
+            if (dims not in self.measure_names) and not self.all_measurements:
                 continue
 
-            ent.label_ = self.measure_names[dims]
+            if self.all_measurements:
+                if not Span.has_extension(unit_norm):
+                    Span.set_extension(unit_norm, default=None)
+                ent.label_ = unit_norm
+
+            else:
+                ent.label_ = self.measure_names[dims]
             ent._.set(
-                self.measure_names[dims],
+                ent.label_,
                 SimpleMeasurement(value, unit_norm, self.unit_registry)
             )